Implemented stack/write code patches for x86-64 and fixed an associated bug in trees.c.
git-svn-id: https://dosemu.svn.sourceforge.net/svnroot/dosemu/trunk@1657 7758b857-130e-0410-ab9c-d8735a528d18

+++ src/emu-i386/simx86/trees.c
@@ -1158,7 +1158,7 @@ static void BreakNode(TNode *G, long eip, long addr)
 	p = G->addr + A->daddr;		// translated IP of following instr
 	if (enpc >= A->dnpc) {		// if it's a forward write
 	    memcpy(p, TailCode, TAILSIZE);
-	    *((long *)(p+TAILFIX)) = G->key + A->dnpc;
+	    *((int *)(p+TAILFIX)) = G->key + A->dnpc;
 	    dbug_printf("============ Force node closing at %08lx(%08lx)\n",
 		(G->key+A->dnpc),(long)p);
 	}
